Initial commit
This repo should be unpacked into the `.ssh` directory of a Linux user.
That user will then be responsible for running the _MapTool_ server and
configuring SSH public keys to allow multiple GMs and/or players to
connect at once.  (SSH public keys are used in addition to the _MapTool_
password to provide fine-grained control.)