Converted 'activate' script to Python
The new script searches ~maptool/Downloads to find the most recent *.txt
file and treats it as a CSV of players/GMs and their keys.  This allows
some managing the server to open Firefox, navigate to a GDrive location,
then click to download a file.  A GDoc form can be used to populate said
file.

This script also allows administrative keys to be stored in *.txt files
under the ~maptool/.ssh/admin-keys directory.  They are one key per line
and as many lines as desired.  This makes it easy to accept a key
uploaded directly to this location (such as by copy/pasting a key from
PuTTY into a Notepad file, then uploading that file).